refactor: :recycle: use composable to handle csvapi calls and variable changes
