+++ udata-tabular-preview/csvapi.js
@@ -1,5 +1,4 @@
 import { apify, configure, getData } from "@etalab/explore.data.gouv.fr/lib/csvapi";
-import { tabular_csvapi_url, tabular_page_size } from "./config";
 
 /**
  * @type {Map<string, Promise<import("@etalab/explore.data.gouv.fr/lib/csvapi").CsvapiResponse>>}
@@ -7,21 +6,23 @@ import { tabular_csvapi_url, tabular_page_size } from "./config";
 const csvapiRequests = new Map();
 
 /**
+ * Call Csvapi functions apify and getData with provided configuration.
+ * Configuration is called twice : one time before each call to Csvapi.
  *
- * @param {{url: string}} resource - a resource hosted on udata
+ * @param {string} url - a url to a resource hosted on udata
+ * @param {import("@etalab/explore.data.gouv.fr/lib/csvapi").CsvapiRequestConfiguration} config - Csvapi configuration
  * @returns {Promise<import("@etalab/explore.data.gouv.fr/lib/csvapi").CsvapiResponse>}
  */
-export default function requestCsvapi({url}) {
-  configure({ csvapiUrl: tabular_csvapi_url, pageSize: tabular_page_size });
-
+export default function requestCsvapi(url, config) {
   let existingRequest;
   if(existingRequest = csvapiRequests.get(url)) {
     return existingRequest;
   }
-
+  configure(config);
   const request = apify(url).then(res => {
     if (res.ok) {
-      configure({ dataEndpoint: res.endpoint });
+      config.dataEndpoint = res.endpoint;
+      configure(config);
       return getData("apify");
     } else {
       throw new Error("Got 200 but result isn't ok");
