+++ udata-tabular-preview/data-structure.vue
@@ -20,37 +20,30 @@
 import { defineComponent, ref } from 'vue';
 import requestCsvapi from './csvapi';
 import Loader from "./loader.vue";
+import useCsvapi from './useCsvapi';
 
 export default defineComponent({
   components: {Loader},
   props: {
     resource: {
-      /** @type {import("vue").PropType<{url: string}>} */
+      /** @type {import("vue").PropType<import("./index").Resource>} */
       type: Object,
       required: true
     }
   },
   setup(props) {
-    /** @type {import("vue").Ref<Array>} */
-    const columns = ref([]);
-    const columnsInfos = ref({});
-    const loading = ref(true);
-    const hasError = ref(false);
-
-    requestCsvapi(props.resource).then(res => {
-          if (res.ok) {
-            columns.value = res.columns;
-            columnsInfos.value = res.columns_infos;
-          } else {
-            hasError.value = true;
-          }
-        }).catch(() => hasError.value = true)
-        .finally(() => loading.value = false);
-    return {
-      hasError,
-      loading,
+    const {
       columns,
       columnsInfos,
+      hasError,
+      loading,
+    } = useCsvapi(props.resource);
+
+    return {
+      columns,
+      columnsInfos,
+      hasError,
+      loading,
     }
   }
 });
